Moved plotting from head_service.
