+++ tvb_epilepsy/scripts/hypothesis_scripts.py
@@ -8,7 +8,6 @@ from tvb_epilepsy.base.utils.log_error_utils import initialize_logger
 from tvb_epilepsy.base.model.disease_hypothesis import DiseaseHypothesis
 from tvb_epilepsy.io.h5_writer import H5Writer
 from tvb_epilepsy.plot.plotter import Plotter
-from tvb_epilepsy.service.head_service import HeadService
 from tvb_epilepsy.service.model_configuration_service import ModelConfigurationService
 from tvb_epilepsy.service.lsa_service import LSAService
 
@@ -44,8 +43,8 @@ def from_head_to_hypotheses(ep_name, data_mode=DATA_MODE, data_folder=os.path.jo
     logger.info("Reading from: " + data_folder)
     head = reader.read_head(data_folder)
     if plot_head:
-        head_service = HeadService()
-        head_service.plot_head(head, figure_dir=figure_dir)
+        plotter = Plotter()
+        plotter.plot_head(head, figure_dir=figure_dir)
         # head.plot(figure_dir=figure_dir)
     # --------------------------Hypothesis definition-----------------------------------
     # # Manual definition of hypothesis...:
